requirements.txt
langchain 
langchain-community 
# anthropic 
fastapi 
uvicorn 
python-dotenv 
requests
openai